MTL-1482 Replace missed variable for bios-hpe-baseline.sh
The variable for the ini file was not being used, this was missed in
testing. The testing was successful in the BMC application but only
because it ran within the same directory as the typoed file path. This
also removes an alias for ilorest, it was not working properly in each
function so now each call to ilorest adds --nologo.