MTL-1876 Use `rootfs`
To align to what CFS uses, and to make the entire boot process the same
for bootstrap, runtime, and customized images (CFS/IMS), we'll use
`rootfs` as the filename instead of `filesystem.squashfs`. This will
prevent us from needing to fiddle the `rd.live.squashimg` parameter at
various points of an install or upgrade.