MTL-2080 Ignore `sha256sum` `initrd` files
The `rootfs` and `kernel` files search by suffix and in doing so they do
not accidentally select `sha256` sum files.

The `initrd` searches by prefix because the suffix could change in the
future (if we change compression or format). Since it searches by prefix
the search picks up the `sha256` sums. This change will ensure those are
excluded.