MTL-2000 Invoke `csi patch packages` (#70)
MTL-2000 Invoke `csi patch packages` and use noos

Publish to noos in addition to SP4, SP3, and SP2.

`csi patch packages` is a new command from MTL-2000 that will patch a
given `cloud-init.yaml` file into `data.json`. The intention here is
that `cloud-init.yaml` contains a list of `repos` and `packages` that we
want cloud-init to install when it runs on the deployed nodes.

If the `cloud-init.yaml` file does not exist in the tar ball, and none
is provided on the command line, then this step is skipped.
Additionally, if `csi patch packages` is not available then the step is
skipped.

Lastly this adds some skip options and a usage for `pit-init.sh`.